requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-loganalytics==13.0.0b6
azure-mgmt-securityinsight==1.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
isodate==0.6.1
lxml==4.9.2
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-docx==0.8.11
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.2
